+++ transforms/no-implicit-this/helpers/known-helpers.js
@@ -8,16 +8,23 @@ const KNOWN_HELPERS = [
   'each',
   'each-in',
   'else',
+  'fn',
   'get',
   'hash',
+  'has-block',
+  'hasBlock',
   'if',
   'if-unless',
   'in-element',
   '-in-element',
+  'input',
+  'textarea',
   'let',
+  'link-to',
   'loc',
   'log',
   'mut',
+  'on',
   'outlet',
   'partial',
   'query-params',
