+++ transforms/no-implicit-this/test.js
@@ -1,21 +1,9 @@
 'use strict';
 
-const path = require('path');
 const { runTransformTest } = require('codemod-cli');
-const { setTelemetry } = require('ember-codemods-telemetry-helpers');
+const { setupTelemetry } = require('./test-helpers');
 
-const mockTelemetryData = require('./__testfixtures__/-mock-telemetry.json');
-
-let mockTelemetry = {};
-
-Object.keys(mockTelemetryData).forEach(key => {
-  let value = mockTelemetryData[key] || {};
-  let mockPath = path.resolve(__dirname, `./__testfixtures__/${key}`);
-
-  mockTelemetry[mockPath] = value;
-});
-
-setTelemetry(mockTelemetry);
+setupTelemetry();
 
 runTransformTest({
   type: 'jscodeshift',
